+++ src/pages/TrackingPage/FurnaceTracking.jsx
@@ -2,19 +2,21 @@ import { PageLayout } from '../pages';
 import '../../styles/_styles.scss';
 
 // CUSTOM COMPONENTS
-import TrackingFormFurnace from '../../components/TrackingForm/TrackingFurnace';
+import TrackingFormFurnace from '../../components/TrackingForm/TrackingFurnaceForm';
 
 export default function FurnaceIntakePage() {
   return (
-    <PageLayout>
-      <div className="header-three jc-center">Intake Screen</div>
-      <div className="">
-        <div className="header-two jc-center">Furnace</div>
-        <div className="header-three jc-center">Maintance Tracking</div>
-        <TrackingFormFurnace />
-        <div className="row jc-center">
+    <div className="login-container">
+        <PageLayout>
+        <div className="header-three jc-center">Intake Screen</div>
+        <div className="">
+            <div className="header-two jc-center">Furnace</div>
+            <div className="header-three jc-center">Maintance Tracking</div>
+            <TrackingFormFurnace />
+            <div className="row jc-center">
+            </div>
         </div>
-      </div>
-    </PageLayout>
+        </PageLayout>
+    </div>
   );
 }
