+++ src/pages/TrackingPage/_trackingPage.scss
@@ -1,0 +1,11 @@
+.tracking {
+    &-container {
+      margin-top: 4em;
+      .btn_asLink {
+        border: none;
+      }
+      .dropdown-menu.jc-end {
+        display: none;
+      }
+    }
+  }
